+++ Alloy/commands/compile/sourceMapper.js
@@ -106,7 +106,8 @@ exports.generateCodeAndSourceMap = function(generator, compileConfig) {
 	try {
 		ast = babylon.parse(genMap.code, {
 			sourceFilename: outfile,
-			sourceType: 'module'
+			sourceType: 'unambiguous',
+			allowReturnOutsideFunction: true
 		});
 	} catch (e) {
 		let filename;
@@ -211,7 +212,8 @@ exports.generateSourceMap = function(generator, compileConfig) {
 	try {
 		ast = babylon.parse(genMap.code, {
 			sourceFilename: genMap.file,
-			sourceType: 'module'
+			sourceType: 'unambiguous',
+			allowReturnOutsideFunction: true,
 		});
 	} catch (e) {
 		const filename = path.relative(compileConfig.dir.project, generator.target.template);